+++ datastructures/Set.h
@@ -56,7 +56,7 @@ Set *BDB_Btree_Set_create(const char *filepath);
 Set *BDB_Hash_Set_create(const char *filepath);
 
 /* Creates a set data structure backed by a custom B-tree implementation. */
-Set *Btree_Set_create(const char *filepath, size_t pagesize);
+Set *Btree_Set_create(const char *filepath, int pagesize);
 
 /* Creates a set data structure backed by a custom hash table implementation. */
 Set *Hash_Set_create(const char *filepath, size_t capacity);
